Add auth based on role

+++ 6. Backend Beginner/tickitz-moviez/src/controllers/users.js
@@ -24,9 +24,9 @@ ctrl.addUser = async (req, res) => {
   }
 }
 
-ctrl.getAllUsers = async (req, res) => {
+ctrl.getUserInfo = async (req, res) => {
   try {
-    const data = await model.getAllUsers()
+    const data = await model.getUserInfo(req.decode)
     return response(res, 200, data)
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
@@ -54,8 +54,12 @@ ctrl.changePassword = async (req, res) => {
 
 ctrl.deleteUser = async (req, res) => {
   try {
-    const data = await model.deleteUser(req.params)
-    return response(res, 200, data)
+    if (req.decode.role === 'admin') {
+      const data = await model.deleteUser(req.params)
+      return response(res, 200, data)
+    } else {
+      return response(res, 401, 'Anda tidak memiliki akses', true)
+    }
   } catch (error) {
     return response(res, 500, 'Terjadi kesalahan', true)
   }
